// src/utils/enums/CardEnum.ts
export enum PerformanceCardType {
    LOW_SCORE = "low_performance",
    MEDIUM_SCORE = "medium_performance",
    HIGH_SCORE = "high_performance",
}

export enum BaseCardType {
    NORMAL = "normal",
    COMPACT = "compact",
}

export enum OverviewType {
    VESSEL_OVERVIEW = "vessel_overview",
    PORT_OVERVIEW = "port_overview"
}